pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-ear"
version = "0.1.0"
description = "Enterprise-grade multi-modal AI audio listening and understanding system"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["ai", "audio", "speech", "nlp", "enterprise", "real-time"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Audio I/O and processing
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "numpy>=1.26.0",
    "librosa>=0.10.1",
    # Speech recognition
    "openai-whisper>=20231117",
    # Transformers / emotion / classification
    "transformers>=4.38.0",
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
    # API / streaming
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "python-multipart>=0.0.9",
    # Utilities
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "structlog>=24.1.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[project.scripts]
ai-ear = "ai_ear.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_ear*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
